requirements.txt
# Core Web Framework
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# AI/ML Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.0
langchain-openai>=0.0.0
langgraph>=0.0.0
openai>=1.0.0

# Database
psycopg2-binary>=2.9.0
pgvector>=0.2.0
redis>=5.0.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0
greenlet>=3.0.0

# Document Processing
pymupdf>=1.24.0
unstructured>=0.13.0
rank-bm25>=0.2.0
markdown>=3.0

# Embedding Models
transformers>=4.38.0
sentence-transformers>=2.6.0
torch>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
aiofiles>=23.2.0
tenacity>=8.2.0

# Frontend
streamlit>=1.32.0
streamlit-webrtc>=0.45.0

# Multimodal Support
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.26.0
pydub>=0.25.1
edge-tts>=6.0.0

# Development Tools
pytest>=8.1.0
pytest-asyncio>=0.23.0
httpx>=0.27.0